fix siren and other sounds?

+++ GameGroup8/Assets/Scripts/Enemy/EnemySpawner.cs
@@ -348,7 +348,7 @@ public class EnemySpawner : MonoBehaviour {
 	/// calculates the amount of enemies that have to be spawned the next wave. It gets higher if the wave is higher.
 	/// </summary>
     void setEnemiesThisWave(){
-		enemiesThisWave = Mathf.Ceil(25f * (2f / Mathf.PI) * Mathf.Atan ((float)wave / 8f));
+		enemiesThisWave = (int)Mathf.Ceil(25f * (2f / Mathf.PI) * Mathf.Atan ((float)wave / 8f));
 		Debug.Log (enemiesThisWave);
 	}
